Pass OPRF encoded inputs with hasher via struct
Rather than passing each encoded input into a buffered channel and
the hasher via it's own channel, we combine them into a new struct.
The struct contains the full slice of OPRF encoded inputs along
with the hasher. Rather than using an error channel, we panic as
the only possible error is due to programmer error in generating
the seed for hashing or AES encoding.