Property-based testing of matrix transposition (#65)
* Add property-based testing of matrix transpose

Rather than just testing the self-inverse property of matrix
transposition, use property-based testing to check if transposition
is correct and meets the self-inverse (involutory) property.

Similarly test unraveling and reraveling of matrices.

* Matrices use randomness from Quick testing methods

We take the default time-seeded randomness from Quick and use it
when generating our matrices. In addition, matrices are generated
with a random length such that the max size is 32 blocks of 512 by
512 bits.

* Fix property-based testing of bit utils

Based on review comments on property-based on testing of matrix
transposition, realized that the testing of the bit utils had
similar flaws. Ensure the randomness from Quick (time-seeded) is
passed to bitset generation and also create sets of truly random
length.